# public/swagger/data/swagger-data.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manager Pokemons",
        "description": "Complete crud for pokemons",
        "version": "1.0.0"
    },
    "paths": {
        "/api/pokemons": {
            "get": {
                "tags": [
                    "pokemons"
                ],
                "summary": "Get all pokemons",
                "operationId": "92dd3cd96c7e01f44cc0ce49d8616c01",
                "responses": {
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Pokemon"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerPokemonAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "pokemons"
                ],
                "summary": "Create a new Pokemon",
                "operationId": "2eac4db9adc623aefa7801dabd6f05fd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "color": {
                                        "type": "string"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "format": "url"
                                    },
                                    "level": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Pokemon created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pokemon"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerPokemonAuth": []
                    }
                ]
            }
        },
        "/api/pokemons/{id}": {
            "get": {
                "tags": [
                    "pokemons"
                ],
                "summary": "Get a specific pokemon by ID",
                "operationId": "10c1548d2490cc2df66b03360afe3ab7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the pokemon to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "User not found or Pokemon not found"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pokemon"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerPokemonAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "pokemons"
                ],
                "summary": "Update a Pokemon",
                "operationId": "e912e4dd886610db61014c84e3313308",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Pokemon to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "color": {
                                        "type": "string"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "format": "url"
                                    },
                                    "level": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Pokemon not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Pokemon updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerPokemonAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "pokemons"
                ],
                "summary": "Delete a Pokemon",
                "operationId": "12647bf2b571550c86bb6a672e965b48",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Pokemon to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Pokemon not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Pokemon deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerPokemonAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Create new user and return your bearer auth",
                "operationId": "7c2f70a3743c58ead609cd340889444b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ok. Return Bearer unique token"
                    },
                    "422": {
                        "description": "Email already exists"
                    },
                    "500": {
                        "description": "Error during create new user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Pokemon": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "color": {
                        "type": "string"
                    },
                    "image_url": {
                        "type": "string",
                        "format": "url"
                    },
                    "level": {
                        "type": "integer",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerPokemonAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}